+++ changelog.md
@@ -22,6 +22,7 @@
 - Fix missing item model for the radio tower (BluSunrize)
 - Fix firework cartridges crashing when crafted (BluSunrize)
 - Fix firework cartridges not accessing vanilla components (BluSunrize)
+- Fix arc recycling recipes not having their own JEI tab (BluSunrize)
 - Translations Added/Updated: zh_cn (Cactusstudent, SlimeSB, mc-kaishixiaxue, JustAlkaid), tr_tr (RuyaSavascisi)
 
 ##### Version 1.21.1-12.0.0-182